refactor: remove IKMSGenerationMigration
