+++ host-contracts/docs/contract_selectors.txt
@@ -1250,23 +1250,6 @@ IKMSGeneration
 ╰----------+-------------------------------------------------------+--------------------------------------------------------------------╯
 
 
-IKMSGenerationMigration
-
-╭----------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------╮
-| Type     | Signature                                                                                                                                                                          | Selector   |
-+============================================================================================================================================================================================================+
-| Function | initializeFromMigration((uint256,uint256,uint256,uint256,uint256,uint256,(uint8,bytes)[],bytes,address[],bytes32,address[],bytes32,address[],bytes32,uint256,uint8,uint8,uint256)) | 0xa0079e0f |
-|----------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------|
-| Error    | InvalidMigrationConsensusState(uint256)                                                                                                                                            | 0x4502cbf1 |
-|----------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------|
-| Error    | InvalidMigrationCounterState()                                                                                                                                                     | 0xa4d3d4f2 |
-|----------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------|
-| Error    | InvalidMigrationMaterial(uint256)                                                                                                                                                  | 0x16bbaf8d |
-|----------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------|
-| Error    | UnknownMigrationConsensusTxSender(uint256,address)                                                                                                                                 | 0x8bd03097 |
-╰----------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------╯
-
-
 IPauserSet
 
 ╭----------+-------------------------------+--------------------------------------------------------------------╮
