--- host-contracts/rust_bindings/src/ikms_generation_migration.rs
@@ -1,2573 +1,0 @@
-///Module containing a contract's types and functions.
-/**
-
-```solidity
-library IKMSGeneration {
-    type KeyType is uint8;
-    type ParamsType is uint8;
-    struct KeyDigest { KeyType keyType; bytes digest; }
-}
-```*/
-#[allow(
-    non_camel_case_types,
-    non_snake_case,
-    clippy::pub_underscore_fields,
-    clippy::style,
-    clippy::empty_structs_with_brackets
-)]
-pub mod IKMSGeneration {
-    use super::*;
-    use alloy::sol_types as alloy_sol_types;
-    #[derive(serde::Serialize, serde::Deserialize)]
-    #[derive(Default, Debug, PartialEq, Eq, Hash)]
-    #[allow(non_camel_case_types, non_snake_case, clippy::pub_underscore_fields)]
-    #[derive(Clone)]
-    pub struct KeyType(u8);
-    const _: () = {
-        use alloy::sol_types as alloy_sol_types;
-        #[automatically_derived]
-        impl alloy_sol_types::private::SolTypeValue<KeyType> for u8 {
-            #[inline]
-            fn stv_to_tokens(
-                &self,
-            ) -> <alloy::sol_types::sol_data::Uint<
-                8,
-            > as alloy_sol_types::SolType>::Token<'_> {
-                alloy_sol_types::private::SolTypeValue::<
-                    alloy::sol_types::sol_data::Uint<8>,
-                >::stv_to_tokens(self)
-            }
-            #[inline]
-            fn stv_eip712_data_word(&self) -> alloy_sol_types::Word {
-                <alloy::sol_types::sol_data::Uint<
-                    8,
-                > as alloy_sol_types::SolType>::tokenize(self)
-                    .0
-            }
-            #[inline]
-            fn stv_abi_encode_packed_to(
-                &self,
-                out: &mut alloy_sol_types::private::Vec<u8>,
-            ) {
-                <alloy::sol_types::sol_data::Uint<
-                    8,
-                > as alloy_sol_types::SolType>::abi_encode_packed_to(self, out)
-            }
-            #[inline]
-            fn stv_abi_packed_encoded_size(&self) -> usize {
-                <alloy::sol_types::sol_data::Uint<
-                    8,
-                > as alloy_sol_types::SolType>::abi_encoded_size(self)
-            }
-        }
-        #[automatically_derived]
-        impl KeyType {
-            /// The Solidity type name.
-            pub const NAME: &'static str = stringify!(@ name);
-            /// Convert from the underlying value type.
-            #[inline]
-            pub const fn from_underlying(value: u8) -> Self {
-                Self(value)
-            }
-            /// Return the underlying value.
-            #[inline]
-            pub const fn into_underlying(self) -> u8 {
-                self.0
-            }
-            /// Return the single encoding of this value, delegating to the
-            /// underlying type.
-            #[inline]
-            pub fn abi_encode(&self) -> alloy_sol_types::private::Vec<u8> {
-                <Self as alloy_sol_types::SolType>::abi_encode(&self.0)
-            }
-            /// Return the packed encoding of this value, delegating to the
-            /// underlying type.
-            #[inline]
-            pub fn abi_encode_packed(&self) -> alloy_sol_types::private::Vec<u8> {
-                <Self as alloy_sol_types::SolType>::abi_encode_packed(&self.0)
-            }
-        }
-        #[automatically_derived]
-        impl From<u8> for KeyType {
-            fn from(value: u8) -> Self {
-                Self::from_underlying(value)
-            }
-        }
-        #[automatically_derived]
-        impl From<KeyType> for u8 {
-            fn from(value: KeyType) -> Self {
-                value.into_underlying()
-            }
-        }
-        #[automatically_derived]
-        impl alloy_sol_types::SolType for KeyType {
-            type RustType = u8;
-            type Token<'a> = <alloy::sol_types::sol_data::Uint<
-                8,
-            > as alloy_sol_types::SolType>::Token<'a>;
-            const SOL_NAME: &'static str = Self::NAME;
-            const ENCODED_SIZE: Option<usize> = <alloy::sol_types::sol_data::Uint<
-                8,
-            > as alloy_sol_types::SolType>::ENCODED_SIZE;
-            const PACKED_ENCODED_SIZE: Option<usize> = <alloy::sol_types::sol_data::Uint<
-                8,
-            > as alloy_sol_types::SolType>::PACKED_ENCODED_SIZE;
-            #[inline]
-            fn valid_token(token: &Self::Token<'_>) -> bool {
-                Self::type_check(token).is_ok()
-            }
-            #[inline]
-            fn type_check(token: &Self::Token<'_>) -> alloy_sol_types::Result<()> {
-                <alloy::sol_types::sol_data::Uint<
-                    8,
-                > as alloy_sol_types::SolType>::type_check(token)
-            }
-            #[inline]
-            fn detokenize(token: Self::Token<'_>) -> Self::RustType {
-                <alloy::sol_types::sol_data::Uint<
-                    8,
-                > as alloy_sol_types::SolType>::detokenize(token)
-            }
-        }
-        #[automatically_derived]
-        impl alloy_sol_types::EventTopic for KeyType {
-            #[inline]
-            fn topic_preimage_length(rust: &Self::RustType) -> usize {
-                <alloy::sol_types::sol_data::Uint<
-                    8,
-                > as alloy_sol_types::EventTopic>::topic_preimage_length(rust)
-            }
-            #[inline]
-            fn encode_topic_preimage(
-                rust: &Self::RustType,
-                out: &mut alloy_sol_types::private::Vec<u8>,
-            ) {
-                <alloy::sol_types::sol_data::Uint<
-                    8,
-                > as alloy_sol_types::EventTopic>::encode_topic_preimage(rust, out)
-            }
-            #[inline]
-            fn encode_topic(
-                rust: &Self::RustType,
-            ) -> alloy_sol_types::abi::token::WordToken {
-                <alloy::sol_types::sol_data::Uint<
-                    8,
-                > as alloy_sol_types::EventTopic>::encode_topic(rust)
-            }
-        }
-    };
-    #[derive(serde::Serialize, serde::Deserialize)]
-    #[derive(Default, Debug, PartialEq, Eq, Hash)]
-    #[allow(non_camel_case_types, non_snake_case, clippy::pub_underscore_fields)]
-    #[derive(Clone)]
-    pub struct ParamsType(u8);
-    const _: () = {
-        use alloy::sol_types as alloy_sol_types;
-        #[automatically_derived]
-        impl alloy_sol_types::private::SolTypeValue<ParamsType> for u8 {
-            #[inline]
-            fn stv_to_tokens(
-                &self,
-            ) -> <alloy::sol_types::sol_data::Uint<
-                8,
-            > as alloy_sol_types::SolType>::Token<'_> {
-                alloy_sol_types::private::SolTypeValue::<
-                    alloy::sol_types::sol_data::Uint<8>,
-                >::stv_to_tokens(self)
-            }
-            #[inline]
-            fn stv_eip712_data_word(&self) -> alloy_sol_types::Word {
-                <alloy::sol_types::sol_data::Uint<
-                    8,
-                > as alloy_sol_types::SolType>::tokenize(self)
-                    .0
-            }
-            #[inline]
-            fn stv_abi_encode_packed_to(
-                &self,
-                out: &mut alloy_sol_types::private::Vec<u8>,
-            ) {
-                <alloy::sol_types::sol_data::Uint<
-                    8,
-                > as alloy_sol_types::SolType>::abi_encode_packed_to(self, out)
-            }
-            #[inline]
-            fn stv_abi_packed_encoded_size(&self) -> usize {
-                <alloy::sol_types::sol_data::Uint<
-                    8,
-                > as alloy_sol_types::SolType>::abi_encoded_size(self)
-            }
-        }
-        #[automatically_derived]
-        impl ParamsType {
-            /// The Solidity type name.
-            pub const NAME: &'static str = stringify!(@ name);
-            /// Convert from the underlying value type.
-            #[inline]
-            pub const fn from_underlying(value: u8) -> Self {
-                Self(value)
-            }
-            /// Return the underlying value.
-            #[inline]
-            pub const fn into_underlying(self) -> u8 {
-                self.0
-            }
-            /// Return the single encoding of this value, delegating to the
-            /// underlying type.
-            #[inline]
-            pub fn abi_encode(&self) -> alloy_sol_types::private::Vec<u8> {
-                <Self as alloy_sol_types::SolType>::abi_encode(&self.0)
-            }
-            /// Return the packed encoding of this value, delegating to the
-            /// underlying type.
-            #[inline]
-            pub fn abi_encode_packed(&self) -> alloy_sol_types::private::Vec<u8> {
-                <Self as alloy_sol_types::SolType>::abi_encode_packed(&self.0)
-            }
-        }
-        #[automatically_derived]
-        impl From<u8> for ParamsType {
-            fn from(value: u8) -> Self {
-                Self::from_underlying(value)
-            }
-        }
-        #[automatically_derived]
-        impl From<ParamsType> for u8 {
-            fn from(value: ParamsType) -> Self {
-                value.into_underlying()
-            }
-        }
-        #[automatically_derived]
-        impl alloy_sol_types::SolType for ParamsType {
-            type RustType = u8;
-            type Token<'a> = <alloy::sol_types::sol_data::Uint<
-                8,
-            > as alloy_sol_types::SolType>::Token<'a>;
-            const SOL_NAME: &'static str = Self::NAME;
-            const ENCODED_SIZE: Option<usize> = <alloy::sol_types::sol_data::Uint<
-                8,
-            > as alloy_sol_types::SolType>::ENCODED_SIZE;
-            const PACKED_ENCODED_SIZE: Option<usize> = <alloy::sol_types::sol_data::Uint<
-                8,
-            > as alloy_sol_types::SolType>::PACKED_ENCODED_SIZE;
-            #[inline]
-            fn valid_token(token: &Self::Token<'_>) -> bool {
-                Self::type_check(token).is_ok()
-            }
-            #[inline]
-            fn type_check(token: &Self::Token<'_>) -> alloy_sol_types::Result<()> {
-                <alloy::sol_types::sol_data::Uint<
-                    8,
-                > as alloy_sol_types::SolType>::type_check(token)
-            }
-            #[inline]
-            fn detokenize(token: Self::Token<'_>) -> Self::RustType {
-                <alloy::sol_types::sol_data::Uint<
-                    8,
-                > as alloy_sol_types::SolType>::detokenize(token)
-            }
-        }
-        #[automatically_derived]
-        impl alloy_sol_types::EventTopic for ParamsType {
-            #[inline]
-            fn topic_preimage_length(rust: &Self::RustType) -> usize {
-                <alloy::sol_types::sol_data::Uint<
-                    8,
-                > as alloy_sol_types::EventTopic>::topic_preimage_length(rust)
-            }
-            #[inline]
-            fn encode_topic_preimage(
-                rust: &Self::RustType,
-                out: &mut alloy_sol_types::private::Vec<u8>,
-            ) {
-                <alloy::sol_types::sol_data::Uint<
-                    8,
-                > as alloy_sol_types::EventTopic>::encode_topic_preimage(rust, out)
-            }
-            #[inline]
-            fn encode_topic(
-                rust: &Self::RustType,
-            ) -> alloy_sol_types::abi::token::WordToken {
-                <alloy::sol_types::sol_data::Uint<
-                    8,
-                > as alloy_sol_types::EventTopic>::encode_topic(rust)
-            }
-        }
-    };
-    #[derive(serde::Serialize, serde::Deserialize)]
-    #[derive(Default, Debug, PartialEq, Eq, Hash)]
-    /**```solidity
-struct KeyDigest { KeyType keyType; bytes digest; }
-```*/
-    #[allow(non_camel_case_types, non_snake_case, clippy::pub_underscore_fields)]
-    #[derive(Clone)]
-    pub struct KeyDigest {
-        #[allow(missing_docs)]
-        pub keyType: <KeyType as alloy::sol_types::SolType>::RustType,
-        #[allow(missing_docs)]
-        pub digest: alloy::sol_types::private::Bytes,
-    }
-    #[allow(
-        non_camel_case_types,
-        non_snake_case,
-        clippy::pub_underscore_fields,
-        clippy::style
-    )]
-    const _: () = {
-        use alloy::sol_types as alloy_sol_types;
-        #[doc(hidden)]
-        type UnderlyingSolTuple<'a> = (KeyType, alloy::sol_types::sol_data::Bytes);
-        #[doc(hidden)]
-        type UnderlyingRustTuple<'a> = (
-            <KeyType as alloy::sol_types::SolType>::RustType,
-            alloy::sol_types::private::Bytes,
-        );
-        #[cfg(test)]
-        #[allow(dead_code, unreachable_patterns)]
-        fn _type_assertion(
-            _t: alloy_sol_types::private::AssertTypeEq<UnderlyingRustTuple>,
-        ) {
-            match _t {
-                alloy_sol_types::private::AssertTypeEq::<
-                    <UnderlyingSolTuple as alloy_sol_types::SolType>::RustType,
-                >(_) => {}
-            }
-        }
-        #[automatically_derived]
-        #[doc(hidden)]
-        impl ::core::convert::From<KeyDigest> for UnderlyingRustTuple<'_> {
-            fn from(value: KeyDigest) -> Self {
-                (value.keyType, value.digest)
-            }
-        }
-        #[automatically_derived]
-        #[doc(hidden)]
-        impl ::core::convert::From<UnderlyingRustTuple<'_>> for KeyDigest {
-            fn from(tuple: UnderlyingRustTuple<'_>) -> Self {
-                Self {
-                    keyType: tuple.0,
-                    digest: tuple.1,
-                }
-            }
-        }
-        #[automatically_derived]
-        impl alloy_sol_types::SolValue for KeyDigest {
-            type SolType = Self;
-        }
-        #[automatically_derived]
-        impl alloy_sol_types::private::SolTypeValue<Self> for KeyDigest {
-            #[inline]
-            fn stv_to_tokens(&self) -> <Self as alloy_sol_types::SolType>::Token<'_> {
-                (
-                    <KeyType as alloy_sol_types::SolType>::tokenize(&self.keyType),
-                    <alloy::sol_types::sol_data::Bytes as alloy_sol_types::SolType>::tokenize(
-                        &self.digest,
-                    ),
-                )
-            }
-            #[inline]
-            fn stv_abi_encoded_size(&self) -> usize {
-                if let Some(size) = <Self as alloy_sol_types::SolType>::ENCODED_SIZE {
-                    return size;
-                }
-                let tuple = <UnderlyingRustTuple<
-                    '_,
-                > as ::core::convert::From<Self>>::from(self.clone());
-                <UnderlyingSolTuple<
-                    '_,
-                > as alloy_sol_types::SolType>::abi_encoded_size(&tuple)
-            }
-            #[inline]
-            fn stv_eip712_data_word(&self) -> alloy_sol_types::Word {
-                <Self as alloy_sol_types::SolStruct>::eip712_hash_struct(self)
-            }
-            #[inline]
-            fn stv_abi_encode_packed_to(
-                &self,
-                out: &mut alloy_sol_types::private::Vec<u8>,
-            ) {
-                let tuple = <UnderlyingRustTuple<
-                    '_,
-                > as ::core::convert::From<Self>>::from(self.clone());
-                <UnderlyingSolTuple<
-                    '_,
-                > as alloy_sol_types::SolType>::abi_encode_packed_to(&tuple, out)
-            }
-            #[inline]
-            fn stv_abi_packed_encoded_size(&self) -> usize {
-                if let Some(size) = <Self as alloy_sol_types::SolType>::PACKED_ENCODED_SIZE {
-                    return size;
-                }
-                let tuple = <UnderlyingRustTuple<
-                    '_,
-                > as ::core::convert::From<Self>>::from(self.clone());
-                <UnderlyingSolTuple<
-                    '_,
-                > as alloy_sol_types::SolType>::abi_packed_encoded_size(&tuple)
-            }
-        }
-        #[automatically_derived]
-        impl alloy_sol_types::SolType for KeyDigest {
-            type RustType = Self;
-            type Token<'a> = <UnderlyingSolTuple<
-                'a,
-            > as alloy_sol_types::SolType>::Token<'a>;
-            const SOL_NAME: &'static str = <Self as alloy_sol_types::SolStruct>::NAME;
-            const ENCODED_SIZE: Option<usize> = <UnderlyingSolTuple<
-                '_,
-            > as alloy_sol_types::SolType>::ENCODED_SIZE;
-            const PACKED_ENCODED_SIZE: Option<usize> = <UnderlyingSolTuple<
-                '_,
-            > as alloy_sol_types::SolType>::PACKED_ENCODED_SIZE;
-            #[inline]
-            fn valid_token(token: &Self::Token<'_>) -> bool {
-                <UnderlyingSolTuple<'_> as alloy_sol_types::SolType>::valid_token(token)
-            }
-            #[inline]
-            fn detokenize(token: Self::Token<'_>) -> Self::RustType {
-                let tuple = <UnderlyingSolTuple<
-                    '_,
-                > as alloy_sol_types::SolType>::detokenize(token);
-                <Self as ::core::convert::From<UnderlyingRustTuple<'_>>>::from(tuple)
-            }
-        }
-        #[automatically_derived]
-        impl alloy_sol_types::SolStruct for KeyDigest {
-            const NAME: &'static str = "KeyDigest";
-            #[inline]
-            fn eip712_root_type() -> alloy_sol_types::private::Cow<'static, str> {
-                alloy_sol_types::private::Cow::Borrowed(
-                    "KeyDigest(uint8 keyType,bytes digest)",
-                )
-            }
-            #[inline]
-            fn eip712_components() -> alloy_sol_types::private::Vec<
-                alloy_sol_types::private::Cow<'static, str>,
-            > {
-                alloy_sol_types::private::Vec::new()
-            }
-            #[inline]
-            fn eip712_encode_type() -> alloy_sol_types::private::Cow<'static, str> {
-                <Self as alloy_sol_types::SolStruct>::eip712_root_type()
-            }
-            #[inline]
-            fn eip712_encode_data(&self) -> alloy_sol_types::private::Vec<u8> {
-                [
-                    <KeyType as alloy_sol_types::SolType>::eip712_data_word(
-                            &self.keyType,
-                        )
-                        .0,
-                    <alloy::sol_types::sol_data::Bytes as alloy_sol_types::SolType>::eip712_data_word(
-                            &self.digest,
-                        )
-                        .0,
-                ]
-                    .concat()
-            }
-        }
-        #[automatically_derived]
-        impl alloy_sol_types::EventTopic for KeyDigest {
-            #[inline]
-            fn topic_preimage_length(rust: &Self::RustType) -> usize {
-                0usize
-                    + <KeyType as alloy_sol_types::EventTopic>::topic_preimage_length(
-                        &rust.keyType,
-                    )
-                    + <alloy::sol_types::sol_data::Bytes as alloy_sol_types::EventTopic>::topic_preimage_length(
-                        &rust.digest,
-                    )
-            }
-            #[inline]
-            fn encode_topic_preimage(
-                rust: &Self::RustType,
-                out: &mut alloy_sol_types::private::Vec<u8>,
-            ) {
-                out.reserve(
-                    <Self as alloy_sol_types::EventTopic>::topic_preimage_length(rust),
-                );
-                <KeyType as alloy_sol_types::EventTopic>::encode_topic_preimage(
-                    &rust.keyType,
-                    out,
-                );
-                <alloy::sol_types::sol_data::Bytes as alloy_sol_types::EventTopic>::encode_topic_preimage(
-                    &rust.digest,
-                    out,
-                );
-            }
-            #[inline]
-            fn encode_topic(
-                rust: &Self::RustType,
-            ) -> alloy_sol_types::abi::token::WordToken {
-                let mut out = alloy_sol_types::private::Vec::new();
-                <Self as alloy_sol_types::EventTopic>::encode_topic_preimage(
-                    rust,
-                    &mut out,
-                );
-                alloy_sol_types::abi::token::WordToken(
-                    alloy_sol_types::private::keccak256(out),
-                )
-            }
-        }
-    };
-    use alloy::contract as alloy_contract;
-    /**Creates a new wrapper around an on-chain [`IKMSGeneration`](self) contract instance.
-
-See the [wrapper's documentation](`IKMSGenerationInstance`) for more details.*/
-    #[inline]
-    pub const fn new<
-        P: alloy_contract::private::Provider<N>,
-        N: alloy_contract::private::Network,
-    >(
-        address: alloy_sol_types::private::Address,
-        provider: P,
-    ) -> IKMSGenerationInstance<P, N> {
-        IKMSGenerationInstance::<P, N>::new(address, provider)
-    }
-    /**A [`IKMSGeneration`](self) instance.
-
-Contains type-safe methods for interacting with an on-chain instance of the
-[`IKMSGeneration`](self) contract located at a given `address`, using a given
-provider `P`.
-
-If the contract bytecode is available (see the [`sol!`](alloy_sol_types::sol!)
-documentation on how to provide it), the `deploy` and `deploy_builder` methods can
-be used to deploy a new instance of the contract.
-
-See the [module-level documentation](self) for all the available methods.*/
-    #[derive(Clone)]
-    pub struct IKMSGenerationInstance<P, N = alloy_contract::private::Ethereum> {
-        address: alloy_sol_types::private::Address,
-        provider: P,
-        _network: ::core::marker::PhantomData<N>,
-    }
-    #[automatically_derived]
-    impl<P, N> ::core::fmt::Debug for IKMSGenerationInstance<P, N> {
-        #[inline]
-        fn fmt(&self, f: &mut ::core::fmt::Formatter<'_>) -> ::core::fmt::Result {
-            f.debug_tuple("IKMSGenerationInstance").field(&self.address).finish()
-        }
-    }
-    /// Instantiation and getters/setters.
-    #[automatically_derived]
-    impl<
-        P: alloy_contract::private::Provider<N>,
-        N: alloy_contract::private::Network,
-    > IKMSGenerationInstance<P, N> {
-        /**Creates a new wrapper around an on-chain [`IKMSGeneration`](self) contract instance.
-
-See the [wrapper's documentation](`IKMSGenerationInstance`) for more details.*/
-        #[inline]
-        pub const fn new(
-            address: alloy_sol_types::private::Address,
-            provider: P,
-        ) -> Self {
-            Self {
-                address,
-                provider,
-                _network: ::core::marker::PhantomData,
-            }
-        }
-        /// Returns a reference to the address.
-        #[inline]
-        pub const fn address(&self) -> &alloy_sol_types::private::Address {
-            &self.address
-        }
-        /// Sets the address.
-        #[inline]
-        pub fn set_address(&mut self, address: alloy_sol_types::private::Address) {
-            self.address = address;
-        }
-        /// Sets the address and returns `self`.
-        pub fn at(mut self, address: alloy_sol_types::private::Address) -> Self {
-            self.set_address(address);
-            self
-        }
-        /// Returns a reference to the provider.
-        #[inline]
-        pub const fn provider(&self) -> &P {
-            &self.provider
-        }
-    }
-    impl<P: ::core::clone::Clone, N> IKMSGenerationInstance<&P, N> {
-        /// Clones the provider and returns a new instance with the cloned provider.
-        #[inline]
-        pub fn with_cloned_provider(self) -> IKMSGenerationInstance<P, N> {
-            IKMSGenerationInstance {
-                address: self.address,
-                provider: ::core::clone::Clone::clone(&self.provider),
-                _network: ::core::marker::PhantomData,
-            }
-        }
-    }
-    /// Function calls.
-    #[automatically_derived]
-    impl<
-        P: alloy_contract::private::Provider<N>,
-        N: alloy_contract::private::Network,
-    > IKMSGenerationInstance<P, N> {
-        /// Creates a new call builder using this contract instance's provider and address.
-        ///
-        /// Note that the call can be any function call, not just those defined in this
-        /// contract. Prefer using the other methods for building type-safe contract calls.
-        pub fn call_builder<C: alloy_sol_types::SolCall>(
-            &self,
-            call: &C,
-        ) -> alloy_contract::SolCallBuilder<&P, C, N> {
-            alloy_contract::SolCallBuilder::new_sol(&self.provider, &self.address, call)
-        }
-    }
-    /// Event filters.
-    #[automatically_derived]
-    impl<
-        P: alloy_contract::private::Provider<N>,
-        N: alloy_contract::private::Network,
-    > IKMSGenerationInstance<P, N> {
-        /// Creates a new event filter using this contract instance's provider and address.
-        ///
-        /// Note that the type can be any event, not just those defined in this contract.
-        /// Prefer using the other methods for building type-safe event filters.
-        pub fn event_filter<E: alloy_sol_types::SolEvent>(
-            &self,
-        ) -> alloy_contract::Event<&P, E, N> {
-            alloy_contract::Event::new_sol(&self.provider, &self.address)
-        }
-    }
-}
-/**
-
-Generated by the following Solidity interface...
-```solidity
-library IKMSGeneration {
-    type KeyType is uint8;
-    type ParamsType is uint8;
-    struct KeyDigest {
-        KeyType keyType;
-        bytes digest;
-    }
-}
-
-interface IKMSGenerationMigration {
-    struct MigrationState {
-        uint256 prepKeygenCounter;
-        uint256 keyCounter;
-        uint256 crsCounter;
-        uint256 activeKeyId;
-        uint256 activeCrsId;
-        uint256 activePrepKeygenId;
-        IKMSGeneration.KeyDigest[] activeKeyDigests;
-        bytes activeCrsDigest;
-        address[] keyConsensusTxSenders;
-        bytes32 keyConsensusDigest;
-        address[] crsConsensusTxSenders;
-        bytes32 crsConsensusDigest;
-        address[] prepKeygenConsensusTxSenders;
-        bytes32 prepKeygenConsensusDigest;
-        uint256 crsMaxBitLength;
-        IKMSGeneration.ParamsType prepKeygenParamsType;
-        IKMSGeneration.ParamsType crsParamsType;
-        uint256 contextId;
-    }
-
-    error InvalidMigrationConsensusState(uint256 requestId);
-    error InvalidMigrationCounterState();
-    error InvalidMigrationMaterial(uint256 requestId);
-    error UnknownMigrationConsensusTxSender(uint256 requestId, address txSender);
-
-    function initializeFromMigration(MigrationState memory state) external;
-}
-```
-
-...which was generated by the following JSON ABI:
-```json
-[
-  {
-    "type": "function",
-    "name": "initializeFromMigration",
-    "inputs": [
-      {
-        "name": "state",
-        "type": "tuple",
-        "internalType": "struct IKMSGenerationMigration.MigrationState",
-        "components": [
-          {
-            "name": "prepKeygenCounter",
-            "type": "uint256",
-            "internalType": "uint256"
-          },
-          {
-            "name": "keyCounter",
-            "type": "uint256",
-            "internalType": "uint256"
-          },
-          {
-            "name": "crsCounter",
-            "type": "uint256",
-            "internalType": "uint256"
-          },
-          {
-            "name": "activeKeyId",
-            "type": "uint256",
-            "internalType": "uint256"
-          },
-          {
-            "name": "activeCrsId",
-            "type": "uint256",
-            "internalType": "uint256"
-          },
-          {
-            "name": "activePrepKeygenId",
-            "type": "uint256",
-            "internalType": "uint256"
-          },
-          {
-            "name": "activeKeyDigests",
-            "type": "tuple[]",
-            "internalType": "struct IKMSGeneration.KeyDigest[]",
-            "components": [
-              {
-                "name": "keyType",
-                "type": "uint8",
-                "internalType": "enum IKMSGeneration.KeyType"
-              },
-              {
-                "name": "digest",
-                "type": "bytes",
-                "internalType": "bytes"
-              }
-            ]
-          },
-          {
-            "name": "activeCrsDigest",
-            "type": "bytes",
-            "internalType": "bytes"
-          },
-          {
-            "name": "keyConsensusTxSenders",
-            "type": "address[]",
-            "internalType": "address[]"
-          },
-          {
-            "name": "keyConsensusDigest",
-            "type": "bytes32",
-            "internalType": "bytes32"
-          },
-          {
-            "name": "crsConsensusTxSenders",
-            "type": "address[]",
-            "internalType": "address[]"
-          },
-          {
-            "name": "crsConsensusDigest",
-            "type": "bytes32",
-            "internalType": "bytes32"
-          },
-          {
-            "name": "prepKeygenConsensusTxSenders",
-            "type": "address[]",
-            "internalType": "address[]"
-          },
-          {
-            "name": "prepKeygenConsensusDigest",
-            "type": "bytes32",
-            "internalType": "bytes32"
-          },
-          {
-            "name": "crsMaxBitLength",
-            "type": "uint256",
-            "internalType": "uint256"
-          },
-          {
-            "name": "prepKeygenParamsType",
-            "type": "uint8",
-            "internalType": "enum IKMSGeneration.ParamsType"
-          },
-          {
-            "name": "crsParamsType",
-            "type": "uint8",
-            "internalType": "enum IKMSGeneration.ParamsType"
-          },
-          {
-            "name": "contextId",
-            "type": "uint256",
-            "internalType": "uint256"
-          }
-        ]
-      }
-    ],
-    "outputs": [],
-    "stateMutability": "nonpayable"
-  },
-  {
-    "type": "error",
-    "name": "InvalidMigrationConsensusState",
-    "inputs": [
-      {
-        "name": "requestId",
-        "type": "uint256",
-        "internalType": "uint256"
-      }
-    ]
-  },
-  {
-    "type": "error",
-    "name": "InvalidMigrationCounterState",
-    "inputs": []
-  },
-  {
-    "type": "error",
-    "name": "InvalidMigrationMaterial",
-    "inputs": [
-      {
-        "name": "requestId",
-        "type": "uint256",
-        "internalType": "uint256"
-      }
-    ]
-  },
-  {
-    "type": "error",
-    "name": "UnknownMigrationConsensusTxSender",
-    "inputs": [
-      {
-        "name": "requestId",
-        "type": "uint256",
-        "internalType": "uint256"
-      },
-      {
-        "name": "txSender",
-        "type": "address",
-        "internalType": "address"
-      }
-    ]
-  }
-]
-```*/
-#[allow(
-    non_camel_case_types,
-    non_snake_case,
-    clippy::pub_underscore_fields,
-    clippy::style,
-    clippy::empty_structs_with_brackets
-)]
-pub mod IKMSGenerationMigration {
-    use super::*;
-    use alloy::sol_types as alloy_sol_types;
-    /// The creation / init bytecode of the contract.
-    ///
-    /// ```text
-    ///0x
-    /// ```
-    #[rustfmt::skip]
-    #[allow(clippy::all)]
-    pub static BYTECODE: alloy_sol_types::private::Bytes = alloy_sol_types::private::Bytes::from_static(
-        b"",
-    );
-    /// The runtime bytecode of the contract, as deployed on the network.
-    ///
-    /// ```text
-    ///0x
-    /// ```
-    #[rustfmt::skip]
-    #[allow(clippy::all)]
-    pub static DEPLOYED_BYTECODE: alloy_sol_types::private::Bytes = alloy_sol_types::private::Bytes::from_static(
-        b"",
-    );
-    #[derive(serde::Serialize, serde::Deserialize)]
-    #[derive()]
-    /**```solidity
-struct MigrationState { uint256 prepKeygenCounter; uint256 keyCounter; uint256 crsCounter; uint256 activeKeyId; uint256 activeCrsId; uint256 activePrepKeygenId; IKMSGeneration.KeyDigest[] activeKeyDigests; bytes activeCrsDigest; address[] keyConsensusTxSenders; bytes32 keyConsensusDigest; address[] crsConsensusTxSenders; bytes32 crsConsensusDigest; address[] prepKeygenConsensusTxSenders; bytes32 prepKeygenConsensusDigest; uint256 crsMaxBitLength; IKMSGeneration.ParamsType prepKeygenParamsType; IKMSGeneration.ParamsType crsParamsType; uint256 contextId; }
-```*/
-    #[allow(non_camel_case_types, non_snake_case, clippy::pub_underscore_fields)]
-    #[derive(Clone)]
-    pub struct MigrationState {
-        #[allow(missing_docs)]
-        pub prepKeygenCounter: alloy::sol_types::private::primitives::aliases::U256,
-        #[allow(missing_docs)]
-        pub keyCounter: alloy::sol_types::private::primitives::aliases::U256,
-        #[allow(missing_docs)]
-        pub crsCounter: alloy::sol_types::private::primitives::aliases::U256,
-        #[allow(missing_docs)]
-        pub activeKeyId: alloy::sol_types::private::primitives::aliases::U256,
-        #[allow(missing_docs)]
-        pub activeCrsId: alloy::sol_types::private::primitives::aliases::U256,
-        #[allow(missing_docs)]
-        pub activePrepKeygenId: alloy::sol_types::private::primitives::aliases::U256,
-        #[allow(missing_docs)]
-        pub activeKeyDigests: alloy::sol_types::private::Vec<
-            <IKMSGeneration::KeyDigest as alloy::sol_types::SolType>::RustType,
-        >,
-        #[allow(missing_docs)]
-        pub activeCrsDigest: alloy::sol_types::private::Bytes,
-        #[allow(missing_docs)]
-        pub keyConsensusTxSenders: alloy::sol_types::private::Vec<
-            alloy::sol_types::private::Address,
-        >,
-        #[allow(missing_docs)]
-        pub keyConsensusDigest: alloy::sol_types::private::FixedBytes<32>,
-        #[allow(missing_docs)]
-        pub crsConsensusTxSenders: alloy::sol_types::private::Vec<
-            alloy::sol_types::private::Address,
-        >,
-        #[allow(missing_docs)]
-        pub crsConsensusDigest: alloy::sol_types::private::FixedBytes<32>,
-        #[allow(missing_docs)]
-        pub prepKeygenConsensusTxSenders: alloy::sol_types::private::Vec<
-            alloy::sol_types::private::Address,
-        >,
-        #[allow(missing_docs)]
-        pub prepKeygenConsensusDigest: alloy::sol_types::private::FixedBytes<32>,
-        #[allow(missing_docs)]
-        pub crsMaxBitLength: alloy::sol_types::private::primitives::aliases::U256,
-        #[allow(missing_docs)]
-        pub prepKeygenParamsType: <IKMSGeneration::ParamsType as alloy::sol_types::SolType>::RustType,
-        #[allow(missing_docs)]
-        pub crsParamsType: <IKMSGeneration::ParamsType as alloy::sol_types::SolType>::RustType,
-        #[allow(missing_docs)]
-        pub contextId: alloy::sol_types::private::primitives::aliases::U256,
-    }
-    #[allow(
-        non_camel_case_types,
-        non_snake_case,
-        clippy::pub_underscore_fields,
-        clippy::style
-    )]
-    const _: () = {
-        use alloy::sol_types as alloy_sol_types;
-        #[doc(hidden)]
-        type UnderlyingSolTuple<'a> = (
-            alloy::sol_types::sol_data::Uint<256>,
-            alloy::sol_types::sol_data::Uint<256>,
-            alloy::sol_types::sol_data::Uint<256>,
-            alloy::sol_types::sol_data::Uint<256>,
-            alloy::sol_types::sol_data::Uint<256>,
-            alloy::sol_types::sol_data::Uint<256>,
-            alloy::sol_types::sol_data::Array<IKMSGeneration::KeyDigest>,
-            alloy::sol_types::sol_data::Bytes,
-            alloy::sol_types::sol_data::Array<alloy::sol_types::sol_data::Address>,
-            alloy::sol_types::sol_data::FixedBytes<32>,
-            alloy::sol_types::sol_data::Array<alloy::sol_types::sol_data::Address>,
-            alloy::sol_types::sol_data::FixedBytes<32>,
-            alloy::sol_types::sol_data::Array<alloy::sol_types::sol_data::Address>,
-            alloy::sol_types::sol_data::FixedBytes<32>,
-            alloy::sol_types::sol_data::Uint<256>,
-            IKMSGeneration::ParamsType,
-            IKMSGeneration::ParamsType,
-            alloy::sol_types::sol_data::Uint<256>,
-        );
-        #[doc(hidden)]
-        type UnderlyingRustTuple<'a> = (
-            alloy::sol_types::private::primitives::aliases::U256,
-            alloy::sol_types::private::primitives::aliases::U256,
-            alloy::sol_types::private::primitives::aliases::U256,
-            alloy::sol_types::private::primitives::aliases::U256,
-            alloy::sol_types::private::primitives::aliases::U256,
-            alloy::sol_types::private::primitives::aliases::U256,
-            alloy::sol_types::private::Vec<
-                <IKMSGeneration::KeyDigest as alloy::sol_types::SolType>::RustType,
-            >,
-            alloy::sol_types::private::Bytes,
-            alloy::sol_types::private::Vec<alloy::sol_types::private::Address>,
-            alloy::sol_types::private::FixedBytes<32>,
-            alloy::sol_types::private::Vec<alloy::sol_types::private::Address>,
-            alloy::sol_types::private::FixedBytes<32>,
-            alloy::sol_types::private::Vec<alloy::sol_types::private::Address>,
-            alloy::sol_types::private::FixedBytes<32>,
-            alloy::sol_types::private::primitives::aliases::U256,
-            <IKMSGeneration::ParamsType as alloy::sol_types::SolType>::RustType,
-            <IKMSGeneration::ParamsType as alloy::sol_types::SolType>::RustType,
-            alloy::sol_types::private::primitives::aliases::U256,
-        );
-        #[cfg(test)]
-        #[allow(dead_code, unreachable_patterns)]
-        fn _type_assertion(
-            _t: alloy_sol_types::private::AssertTypeEq<UnderlyingRustTuple>,
-        ) {
-            match _t {
-                alloy_sol_types::private::AssertTypeEq::<
-                    <UnderlyingSolTuple as alloy_sol_types::SolType>::RustType,
-                >(_) => {}
-            }
-        }
-        #[automatically_derived]
-        #[doc(hidden)]
-        impl ::core::convert::From<MigrationState> for UnderlyingRustTuple<'_> {
-            fn from(value: MigrationState) -> Self {
-                (
-                    value.prepKeygenCounter,
-                    value.keyCounter,
-                    value.crsCounter,
-                    value.activeKeyId,
-                    value.activeCrsId,
-                    value.activePrepKeygenId,
-                    value.activeKeyDigests,
-                    value.activeCrsDigest,
-                    value.keyConsensusTxSenders,
-                    value.keyConsensusDigest,
-                    value.crsConsensusTxSenders,
-                    value.crsConsensusDigest,
-                    value.prepKeygenConsensusTxSenders,
-                    value.prepKeygenConsensusDigest,
-                    value.crsMaxBitLength,
-                    value.prepKeygenParamsType,
-                    value.crsParamsType,
-                    value.contextId,
-                )
-            }
-        }
-        #[automatically_derived]
-        #[doc(hidden)]
-        impl ::core::convert::From<UnderlyingRustTuple<'_>> for MigrationState {
-            fn from(tuple: UnderlyingRustTuple<'_>) -> Self {
-                Self {
-                    prepKeygenCounter: tuple.0,
-                    keyCounter: tuple.1,
-                    crsCounter: tuple.2,
-                    activeKeyId: tuple.3,
-                    activeCrsId: tuple.4,
-                    activePrepKeygenId: tuple.5,
-                    activeKeyDigests: tuple.6,
-                    activeCrsDigest: tuple.7,
-                    keyConsensusTxSenders: tuple.8,
-                    keyConsensusDigest: tuple.9,
-                    crsConsensusTxSenders: tuple.10,
-                    crsConsensusDigest: tuple.11,
-                    prepKeygenConsensusTxSenders: tuple.12,
-                    prepKeygenConsensusDigest: tuple.13,
-                    crsMaxBitLength: tuple.14,
-                    prepKeygenParamsType: tuple.15,
-                    crsParamsType: tuple.16,
-                    contextId: tuple.17,
-                }
-            }
-        }
-        #[automatically_derived]
-        impl alloy_sol_types::SolValue for MigrationState {
-            type SolType = Self;
-        }
-        #[automatically_derived]
-        impl alloy_sol_types::private::SolTypeValue<Self> for MigrationState {
-            #[inline]
-            fn stv_to_tokens(&self) -> <Self as alloy_sol_types::SolType>::Token<'_> {
-                (
-                    <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::SolType>::tokenize(&self.prepKeygenCounter),
-                    <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::SolType>::tokenize(&self.keyCounter),
-                    <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::SolType>::tokenize(&self.crsCounter),
-                    <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::SolType>::tokenize(&self.activeKeyId),
-                    <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::SolType>::tokenize(&self.activeCrsId),
-                    <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::SolType>::tokenize(&self.activePrepKeygenId),
-                    <alloy::sol_types::sol_data::Array<
-                        IKMSGeneration::KeyDigest,
-                    > as alloy_sol_types::SolType>::tokenize(&self.activeKeyDigests),
-                    <alloy::sol_types::sol_data::Bytes as alloy_sol_types::SolType>::tokenize(
-                        &self.activeCrsDigest,
-                    ),
-                    <alloy::sol_types::sol_data::Array<
-                        alloy::sol_types::sol_data::Address,
-                    > as alloy_sol_types::SolType>::tokenize(
-                        &self.keyConsensusTxSenders,
-                    ),
-                    <alloy::sol_types::sol_data::FixedBytes<
-                        32,
-                    > as alloy_sol_types::SolType>::tokenize(&self.keyConsensusDigest),
-                    <alloy::sol_types::sol_data::Array<
-                        alloy::sol_types::sol_data::Address,
-                    > as alloy_sol_types::SolType>::tokenize(
-                        &self.crsConsensusTxSenders,
-                    ),
-                    <alloy::sol_types::sol_data::FixedBytes<
-                        32,
-                    > as alloy_sol_types::SolType>::tokenize(&self.crsConsensusDigest),
-                    <alloy::sol_types::sol_data::Array<
-                        alloy::sol_types::sol_data::Address,
-                    > as alloy_sol_types::SolType>::tokenize(
-                        &self.prepKeygenConsensusTxSenders,
-                    ),
-                    <alloy::sol_types::sol_data::FixedBytes<
-                        32,
-                    > as alloy_sol_types::SolType>::tokenize(
-                        &self.prepKeygenConsensusDigest,
-                    ),
-                    <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::SolType>::tokenize(&self.crsMaxBitLength),
-                    <IKMSGeneration::ParamsType as alloy_sol_types::SolType>::tokenize(
-                        &self.prepKeygenParamsType,
-                    ),
-                    <IKMSGeneration::ParamsType as alloy_sol_types::SolType>::tokenize(
-                        &self.crsParamsType,
-                    ),
-                    <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::SolType>::tokenize(&self.contextId),
-                )
-            }
-            #[inline]
-            fn stv_abi_encoded_size(&self) -> usize {
-                if let Some(size) = <Self as alloy_sol_types::SolType>::ENCODED_SIZE {
-                    return size;
-                }
-                let tuple = <UnderlyingRustTuple<
-                    '_,
-                > as ::core::convert::From<Self>>::from(self.clone());
-                <UnderlyingSolTuple<
-                    '_,
-                > as alloy_sol_types::SolType>::abi_encoded_size(&tuple)
-            }
-            #[inline]
-            fn stv_eip712_data_word(&self) -> alloy_sol_types::Word {
-                <Self as alloy_sol_types::SolStruct>::eip712_hash_struct(self)
-            }
-            #[inline]
-            fn stv_abi_encode_packed_to(
-                &self,
-                out: &mut alloy_sol_types::private::Vec<u8>,
-            ) {
-                let tuple = <UnderlyingRustTuple<
-                    '_,
-                > as ::core::convert::From<Self>>::from(self.clone());
-                <UnderlyingSolTuple<
-                    '_,
-                > as alloy_sol_types::SolType>::abi_encode_packed_to(&tuple, out)
-            }
-            #[inline]
-            fn stv_abi_packed_encoded_size(&self) -> usize {
-                if let Some(size) = <Self as alloy_sol_types::SolType>::PACKED_ENCODED_SIZE {
-                    return size;
-                }
-                let tuple = <UnderlyingRustTuple<
-                    '_,
-                > as ::core::convert::From<Self>>::from(self.clone());
-                <UnderlyingSolTuple<
-                    '_,
-                > as alloy_sol_types::SolType>::abi_packed_encoded_size(&tuple)
-            }
-        }
-        #[automatically_derived]
-        impl alloy_sol_types::SolType for MigrationState {
-            type RustType = Self;
-            type Token<'a> = <UnderlyingSolTuple<
-                'a,
-            > as alloy_sol_types::SolType>::Token<'a>;
-            const SOL_NAME: &'static str = <Self as alloy_sol_types::SolStruct>::NAME;
-            const ENCODED_SIZE: Option<usize> = <UnderlyingSolTuple<
-                '_,
-            > as alloy_sol_types::SolType>::ENCODED_SIZE;
-            const PACKED_ENCODED_SIZE: Option<usize> = <UnderlyingSolTuple<
-                '_,
-            > as alloy_sol_types::SolType>::PACKED_ENCODED_SIZE;
-            #[inline]
-            fn valid_token(token: &Self::Token<'_>) -> bool {
-                <UnderlyingSolTuple<'_> as alloy_sol_types::SolType>::valid_token(token)
-            }
-            #[inline]
-            fn detokenize(token: Self::Token<'_>) -> Self::RustType {
-                let tuple = <UnderlyingSolTuple<
-                    '_,
-                > as alloy_sol_types::SolType>::detokenize(token);
-                <Self as ::core::convert::From<UnderlyingRustTuple<'_>>>::from(tuple)
-            }
-        }
-        #[automatically_derived]
-        impl alloy_sol_types::SolStruct for MigrationState {
-            const NAME: &'static str = "MigrationState";
-            #[inline]
-            fn eip712_root_type() -> alloy_sol_types::private::Cow<'static, str> {
-                alloy_sol_types::private::Cow::Borrowed(
-                    "MigrationState(uint256 prepKeygenCounter,uint256 keyCounter,uint256 crsCounter,uint256 activeKeyId,uint256 activeCrsId,uint256 activePrepKeygenId,IKMSGeneration.KeyDigest[] activeKeyDigests,bytes activeCrsDigest,address[] keyConsensusTxSenders,bytes32 keyConsensusDigest,address[] crsConsensusTxSenders,bytes32 crsConsensusDigest,address[] prepKeygenConsensusTxSenders,bytes32 prepKeygenConsensusDigest,uint256 crsMaxBitLength,uint8 prepKeygenParamsType,uint8 crsParamsType,uint256 contextId)",
-                )
-            }
-            #[inline]
-            fn eip712_components() -> alloy_sol_types::private::Vec<
-                alloy_sol_types::private::Cow<'static, str>,
-            > {
-                let mut components = alloy_sol_types::private::Vec::with_capacity(1);
-                components
-                    .push(
-                        <IKMSGeneration::KeyDigest as alloy_sol_types::SolStruct>::eip712_root_type(),
-                    );
-                components
-                    .extend(
-                        <IKMSGeneration::KeyDigest as alloy_sol_types::SolStruct>::eip712_components(),
-                    );
-                components
-            }
-            #[inline]
-            fn eip712_encode_data(&self) -> alloy_sol_types::private::Vec<u8> {
-                [
-                    <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::SolType>::eip712_data_word(
-                            &self.prepKeygenCounter,
-                        )
-                        .0,
-                    <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::SolType>::eip712_data_word(&self.keyCounter)
-                        .0,
-                    <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::SolType>::eip712_data_word(&self.crsCounter)
-                        .0,
-                    <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::SolType>::eip712_data_word(&self.activeKeyId)
-                        .0,
-                    <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::SolType>::eip712_data_word(&self.activeCrsId)
-                        .0,
-                    <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::SolType>::eip712_data_word(
-                            &self.activePrepKeygenId,
-                        )
-                        .0,
-                    <alloy::sol_types::sol_data::Array<
-                        IKMSGeneration::KeyDigest,
-                    > as alloy_sol_types::SolType>::eip712_data_word(
-                            &self.activeKeyDigests,
-                        )
-                        .0,
-                    <alloy::sol_types::sol_data::Bytes as alloy_sol_types::SolType>::eip712_data_word(
-                            &self.activeCrsDigest,
-                        )
-                        .0,
-                    <alloy::sol_types::sol_data::Array<
-                        alloy::sol_types::sol_data::Address,
-                    > as alloy_sol_types::SolType>::eip712_data_word(
-                            &self.keyConsensusTxSenders,
-                        )
-                        .0,
-                    <alloy::sol_types::sol_data::FixedBytes<
-                        32,
-                    > as alloy_sol_types::SolType>::eip712_data_word(
-                            &self.keyConsensusDigest,
-                        )
-                        .0,
-                    <alloy::sol_types::sol_data::Array<
-                        alloy::sol_types::sol_data::Address,
-                    > as alloy_sol_types::SolType>::eip712_data_word(
-                            &self.crsConsensusTxSenders,
-                        )
-                        .0,
-                    <alloy::sol_types::sol_data::FixedBytes<
-                        32,
-                    > as alloy_sol_types::SolType>::eip712_data_word(
-                            &self.crsConsensusDigest,
-                        )
-                        .0,
-                    <alloy::sol_types::sol_data::Array<
-                        alloy::sol_types::sol_data::Address,
-                    > as alloy_sol_types::SolType>::eip712_data_word(
-                            &self.prepKeygenConsensusTxSenders,
-                        )
-                        .0,
-                    <alloy::sol_types::sol_data::FixedBytes<
-                        32,
-                    > as alloy_sol_types::SolType>::eip712_data_word(
-                            &self.prepKeygenConsensusDigest,
-                        )
-                        .0,
-                    <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::SolType>::eip712_data_word(
-                            &self.crsMaxBitLength,
-                        )
-                        .0,
-                    <IKMSGeneration::ParamsType as alloy_sol_types::SolType>::eip712_data_word(
-                            &self.prepKeygenParamsType,
-                        )
-                        .0,
-                    <IKMSGeneration::ParamsType as alloy_sol_types::SolType>::eip712_data_word(
-                            &self.crsParamsType,
-                        )
-                        .0,
-                    <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::SolType>::eip712_data_word(&self.contextId)
-                        .0,
-                ]
-                    .concat()
-            }
-        }
-        #[automatically_derived]
-        impl alloy_sol_types::EventTopic for MigrationState {
-            #[inline]
-            fn topic_preimage_length(rust: &Self::RustType) -> usize {
-                0usize
-                    + <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::EventTopic>::topic_preimage_length(
-                        &rust.prepKeygenCounter,
-                    )
-                    + <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::EventTopic>::topic_preimage_length(
-                        &rust.keyCounter,
-                    )
-                    + <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::EventTopic>::topic_preimage_length(
-                        &rust.crsCounter,
-                    )
-                    + <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::EventTopic>::topic_preimage_length(
-                        &rust.activeKeyId,
-                    )
-                    + <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::EventTopic>::topic_preimage_length(
-                        &rust.activeCrsId,
-                    )
-                    + <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::EventTopic>::topic_preimage_length(
-                        &rust.activePrepKeygenId,
-                    )
-                    + <alloy::sol_types::sol_data::Array<
-                        IKMSGeneration::KeyDigest,
-                    > as alloy_sol_types::EventTopic>::topic_preimage_length(
-                        &rust.activeKeyDigests,
-                    )
-                    + <alloy::sol_types::sol_data::Bytes as alloy_sol_types::EventTopic>::topic_preimage_length(
-                        &rust.activeCrsDigest,
-                    )
-                    + <alloy::sol_types::sol_data::Array<
-                        alloy::sol_types::sol_data::Address,
-                    > as alloy_sol_types::EventTopic>::topic_preimage_length(
-                        &rust.keyConsensusTxSenders,
-                    )
-                    + <alloy::sol_types::sol_data::FixedBytes<
-                        32,
-                    > as alloy_sol_types::EventTopic>::topic_preimage_length(
-                        &rust.keyConsensusDigest,
-                    )
-                    + <alloy::sol_types::sol_data::Array<
-                        alloy::sol_types::sol_data::Address,
-                    > as alloy_sol_types::EventTopic>::topic_preimage_length(
-                        &rust.crsConsensusTxSenders,
-                    )
-                    + <alloy::sol_types::sol_data::FixedBytes<
-                        32,
-                    > as alloy_sol_types::EventTopic>::topic_preimage_length(
-                        &rust.crsConsensusDigest,
-                    )
-                    + <alloy::sol_types::sol_data::Array<
-                        alloy::sol_types::sol_data::Address,
-                    > as alloy_sol_types::EventTopic>::topic_preimage_length(
-                        &rust.prepKeygenConsensusTxSenders,
-                    )
-                    + <alloy::sol_types::sol_data::FixedBytes<
-                        32,
-                    > as alloy_sol_types::EventTopic>::topic_preimage_length(
-                        &rust.prepKeygenConsensusDigest,
-                    )
-                    + <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::EventTopic>::topic_preimage_length(
-                        &rust.crsMaxBitLength,
-                    )
-                    + <IKMSGeneration::ParamsType as alloy_sol_types::EventTopic>::topic_preimage_length(
-                        &rust.prepKeygenParamsType,
-                    )
-                    + <IKMSGeneration::ParamsType as alloy_sol_types::EventTopic>::topic_preimage_length(
-                        &rust.crsParamsType,
-                    )
-                    + <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::EventTopic>::topic_preimage_length(
-                        &rust.contextId,
-                    )
-            }
-            #[inline]
-            fn encode_topic_preimage(
-                rust: &Self::RustType,
-                out: &mut alloy_sol_types::private::Vec<u8>,
-            ) {
-                out.reserve(
-                    <Self as alloy_sol_types::EventTopic>::topic_preimage_length(rust),
-                );
-                <alloy::sol_types::sol_data::Uint<
-                    256,
-                > as alloy_sol_types::EventTopic>::encode_topic_preimage(
-                    &rust.prepKeygenCounter,
-                    out,
-                );
-                <alloy::sol_types::sol_data::Uint<
-                    256,
-                > as alloy_sol_types::EventTopic>::encode_topic_preimage(
-                    &rust.keyCounter,
-                    out,
-                );
-                <alloy::sol_types::sol_data::Uint<
-                    256,
-                > as alloy_sol_types::EventTopic>::encode_topic_preimage(
-                    &rust.crsCounter,
-                    out,
-                );
-                <alloy::sol_types::sol_data::Uint<
-                    256,
-                > as alloy_sol_types::EventTopic>::encode_topic_preimage(
-                    &rust.activeKeyId,
-                    out,
-                );
-                <alloy::sol_types::sol_data::Uint<
-                    256,
-                > as alloy_sol_types::EventTopic>::encode_topic_preimage(
-                    &rust.activeCrsId,
-                    out,
-                );
-                <alloy::sol_types::sol_data::Uint<
-                    256,
-                > as alloy_sol_types::EventTopic>::encode_topic_preimage(
-                    &rust.activePrepKeygenId,
-                    out,
-                );
-                <alloy::sol_types::sol_data::Array<
-                    IKMSGeneration::KeyDigest,
-                > as alloy_sol_types::EventTopic>::encode_topic_preimage(
-                    &rust.activeKeyDigests,
-                    out,
-                );
-                <alloy::sol_types::sol_data::Bytes as alloy_sol_types::EventTopic>::encode_topic_preimage(
-                    &rust.activeCrsDigest,
-                    out,
-                );
-                <alloy::sol_types::sol_data::Array<
-                    alloy::sol_types::sol_data::Address,
-                > as alloy_sol_types::EventTopic>::encode_topic_preimage(
-                    &rust.keyConsensusTxSenders,
-                    out,
-                );
-                <alloy::sol_types::sol_data::FixedBytes<
-                    32,
-                > as alloy_sol_types::EventTopic>::encode_topic_preimage(
-                    &rust.keyConsensusDigest,
-                    out,
-                );
-                <alloy::sol_types::sol_data::Array<
-                    alloy::sol_types::sol_data::Address,
-                > as alloy_sol_types::EventTopic>::encode_topic_preimage(
-                    &rust.crsConsensusTxSenders,
-                    out,
-                );
-                <alloy::sol_types::sol_data::FixedBytes<
-                    32,
-                > as alloy_sol_types::EventTopic>::encode_topic_preimage(
-                    &rust.crsConsensusDigest,
-                    out,
-                );
-                <alloy::sol_types::sol_data::Array<
-                    alloy::sol_types::sol_data::Address,
-                > as alloy_sol_types::EventTopic>::encode_topic_preimage(
-                    &rust.prepKeygenConsensusTxSenders,
-                    out,
-                );
-                <alloy::sol_types::sol_data::FixedBytes<
-                    32,
-                > as alloy_sol_types::EventTopic>::encode_topic_preimage(
-                    &rust.prepKeygenConsensusDigest,
-                    out,
-                );
-                <alloy::sol_types::sol_data::Uint<
-                    256,
-                > as alloy_sol_types::EventTopic>::encode_topic_preimage(
-                    &rust.crsMaxBitLength,
-                    out,
-                );
-                <IKMSGeneration::ParamsType as alloy_sol_types::EventTopic>::encode_topic_preimage(
-                    &rust.prepKeygenParamsType,
-                    out,
-                );
-                <IKMSGeneration::ParamsType as alloy_sol_types::EventTopic>::encode_topic_preimage(
-                    &rust.crsParamsType,
-                    out,
-                );
-                <alloy::sol_types::sol_data::Uint<
-                    256,
-                > as alloy_sol_types::EventTopic>::encode_topic_preimage(
-                    &rust.contextId,
-                    out,
-                );
-            }
-            #[inline]
-            fn encode_topic(
-                rust: &Self::RustType,
-            ) -> alloy_sol_types::abi::token::WordToken {
-                let mut out = alloy_sol_types::private::Vec::new();
-                <Self as alloy_sol_types::EventTopic>::encode_topic_preimage(
-                    rust,
-                    &mut out,
-                );
-                alloy_sol_types::abi::token::WordToken(
-                    alloy_sol_types::private::keccak256(out),
-                )
-            }
-        }
-    };
-    #[derive(serde::Serialize, serde::Deserialize)]
-    #[derive(Default, Debug, PartialEq, Eq, Hash)]
-    /**Custom error with signature `InvalidMigrationConsensusState(uint256)` and selector `0x4502cbf1`.
-```solidity
-error InvalidMigrationConsensusState(uint256 requestId);
-```*/
-    #[allow(non_camel_case_types, non_snake_case, clippy::pub_underscore_fields)]
-    #[derive(Clone)]
-    pub struct InvalidMigrationConsensusState {
-        #[allow(missing_docs)]
-        pub requestId: alloy::sol_types::private::primitives::aliases::U256,
-    }
-    #[allow(
-        non_camel_case_types,
-        non_snake_case,
-        clippy::pub_underscore_fields,
-        clippy::style
-    )]
-    const _: () = {
-        use alloy::sol_types as alloy_sol_types;
-        #[doc(hidden)]
-        type UnderlyingSolTuple<'a> = (alloy::sol_types::sol_data::Uint<256>,);
-        #[doc(hidden)]
-        type UnderlyingRustTuple<'a> = (
-            alloy::sol_types::private::primitives::aliases::U256,
-        );
-        #[cfg(test)]
-        #[allow(dead_code, unreachable_patterns)]
-        fn _type_assertion(
-            _t: alloy_sol_types::private::AssertTypeEq<UnderlyingRustTuple>,
-        ) {
-            match _t {
-                alloy_sol_types::private::AssertTypeEq::<
-                    <UnderlyingSolTuple as alloy_sol_types::SolType>::RustType,
-                >(_) => {}
-            }
-        }
-        #[automatically_derived]
-        #[doc(hidden)]
-        impl ::core::convert::From<InvalidMigrationConsensusState>
-        for UnderlyingRustTuple<'_> {
-            fn from(value: InvalidMigrationConsensusState) -> Self {
-                (value.requestId,)
-            }
-        }
-        #[automatically_derived]
-        #[doc(hidden)]
-        impl ::core::convert::From<UnderlyingRustTuple<'_>>
-        for InvalidMigrationConsensusState {
-            fn from(tuple: UnderlyingRustTuple<'_>) -> Self {
-                Self { requestId: tuple.0 }
-            }
-        }
-        #[automatically_derived]
-        impl alloy_sol_types::SolError for InvalidMigrationConsensusState {
-            type Parameters<'a> = UnderlyingSolTuple<'a>;
-            type Token<'a> = <Self::Parameters<
-                'a,
-            > as alloy_sol_types::SolType>::Token<'a>;
-            const SIGNATURE: &'static str = "InvalidMigrationConsensusState(uint256)";
-            const SELECTOR: [u8; 4] = [69u8, 2u8, 203u8, 241u8];
-            #[inline]
-            fn new<'a>(
-                tuple: <Self::Parameters<'a> as alloy_sol_types::SolType>::RustType,
-            ) -> Self {
-                tuple.into()
-            }
-            #[inline]
-            fn tokenize(&self) -> Self::Token<'_> {
-                (
-                    <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::SolType>::tokenize(&self.requestId),
-                )
-            }
-            #[inline]
-            fn abi_decode_raw_validate(data: &[u8]) -> alloy_sol_types::Result<Self> {
-                <Self::Parameters<
-                    '_,
-                > as alloy_sol_types::SolType>::abi_decode_sequence_validate(data)
-                    .map(Self::new)
-            }
-        }
-    };
-    #[derive(serde::Serialize, serde::Deserialize)]
-    #[derive(Default, Debug, PartialEq, Eq, Hash)]
-    /**Custom error with signature `InvalidMigrationCounterState()` and selector `0xa4d3d4f2`.
-```solidity
-error InvalidMigrationCounterState();
-```*/
-    #[allow(non_camel_case_types, non_snake_case, clippy::pub_underscore_fields)]
-    #[derive(Clone)]
-    pub struct InvalidMigrationCounterState;
-    #[allow(
-        non_camel_case_types,
-        non_snake_case,
-        clippy::pub_underscore_fields,
-        clippy::style
-    )]
-    const _: () = {
-        use alloy::sol_types as alloy_sol_types;
-        #[doc(hidden)]
-        type UnderlyingSolTuple<'a> = ();
-        #[doc(hidden)]
-        type UnderlyingRustTuple<'a> = ();
-        #[cfg(test)]
-        #[allow(dead_code, unreachable_patterns)]
-        fn _type_assertion(
-            _t: alloy_sol_types::private::AssertTypeEq<UnderlyingRustTuple>,
-        ) {
-            match _t {
-                alloy_sol_types::private::AssertTypeEq::<
-                    <UnderlyingSolTuple as alloy_sol_types::SolType>::RustType,
-                >(_) => {}
-            }
-        }
-        #[automatically_derived]
-        #[doc(hidden)]
-        impl ::core::convert::From<InvalidMigrationCounterState>
-        for UnderlyingRustTuple<'_> {
-            fn from(value: InvalidMigrationCounterState) -> Self {
-                ()
-            }
-        }
-        #[automatically_derived]
-        #[doc(hidden)]
-        impl ::core::convert::From<UnderlyingRustTuple<'_>>
-        for InvalidMigrationCounterState {
-            fn from(tuple: UnderlyingRustTuple<'_>) -> Self {
-                Self
-            }
-        }
-        #[automatically_derived]
-        impl alloy_sol_types::SolError for InvalidMigrationCounterState {
-            type Parameters<'a> = UnderlyingSolTuple<'a>;
-            type Token<'a> = <Self::Parameters<
-                'a,
-            > as alloy_sol_types::SolType>::Token<'a>;
-            const SIGNATURE: &'static str = "InvalidMigrationCounterState()";
-            const SELECTOR: [u8; 4] = [164u8, 211u8, 212u8, 242u8];
-            #[inline]
-            fn new<'a>(
-                tuple: <Self::Parameters<'a> as alloy_sol_types::SolType>::RustType,
-            ) -> Self {
-                tuple.into()
-            }
-            #[inline]
-            fn tokenize(&self) -> Self::Token<'_> {
-                ()
-            }
-            #[inline]
-            fn abi_decode_raw_validate(data: &[u8]) -> alloy_sol_types::Result<Self> {
-                <Self::Parameters<
-                    '_,
-                > as alloy_sol_types::SolType>::abi_decode_sequence_validate(data)
-                    .map(Self::new)
-            }
-        }
-    };
-    #[derive(serde::Serialize, serde::Deserialize)]
-    #[derive(Default, Debug, PartialEq, Eq, Hash)]
-    /**Custom error with signature `InvalidMigrationMaterial(uint256)` and selector `0x16bbaf8d`.
-```solidity
-error InvalidMigrationMaterial(uint256 requestId);
-```*/
-    #[allow(non_camel_case_types, non_snake_case, clippy::pub_underscore_fields)]
-    #[derive(Clone)]
-    pub struct InvalidMigrationMaterial {
-        #[allow(missing_docs)]
-        pub requestId: alloy::sol_types::private::primitives::aliases::U256,
-    }
-    #[allow(
-        non_camel_case_types,
-        non_snake_case,
-        clippy::pub_underscore_fields,
-        clippy::style
-    )]
-    const _: () = {
-        use alloy::sol_types as alloy_sol_types;
-        #[doc(hidden)]
-        type UnderlyingSolTuple<'a> = (alloy::sol_types::sol_data::Uint<256>,);
-        #[doc(hidden)]
-        type UnderlyingRustTuple<'a> = (
-            alloy::sol_types::private::primitives::aliases::U256,
-        );
-        #[cfg(test)]
-        #[allow(dead_code, unreachable_patterns)]
-        fn _type_assertion(
-            _t: alloy_sol_types::private::AssertTypeEq<UnderlyingRustTuple>,
-        ) {
-            match _t {
-                alloy_sol_types::private::AssertTypeEq::<
-                    <UnderlyingSolTuple as alloy_sol_types::SolType>::RustType,
-                >(_) => {}
-            }
-        }
-        #[automatically_derived]
-        #[doc(hidden)]
-        impl ::core::convert::From<InvalidMigrationMaterial>
-        for UnderlyingRustTuple<'_> {
-            fn from(value: InvalidMigrationMaterial) -> Self {
-                (value.requestId,)
-            }
-        }
-        #[automatically_derived]
-        #[doc(hidden)]
-        impl ::core::convert::From<UnderlyingRustTuple<'_>>
-        for InvalidMigrationMaterial {
-            fn from(tuple: UnderlyingRustTuple<'_>) -> Self {
-                Self { requestId: tuple.0 }
-            }
-        }
-        #[automatically_derived]
-        impl alloy_sol_types::SolError for InvalidMigrationMaterial {
-            type Parameters<'a> = UnderlyingSolTuple<'a>;
-            type Token<'a> = <Self::Parameters<
-                'a,
-            > as alloy_sol_types::SolType>::Token<'a>;
-            const SIGNATURE: &'static str = "InvalidMigrationMaterial(uint256)";
-            const SELECTOR: [u8; 4] = [22u8, 187u8, 175u8, 141u8];
-            #[inline]
-            fn new<'a>(
-                tuple: <Self::Parameters<'a> as alloy_sol_types::SolType>::RustType,
-            ) -> Self {
-                tuple.into()
-            }
-            #[inline]
-            fn tokenize(&self) -> Self::Token<'_> {
-                (
-                    <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::SolType>::tokenize(&self.requestId),
-                )
-            }
-            #[inline]
-            fn abi_decode_raw_validate(data: &[u8]) -> alloy_sol_types::Result<Self> {
-                <Self::Parameters<
-                    '_,
-                > as alloy_sol_types::SolType>::abi_decode_sequence_validate(data)
-                    .map(Self::new)
-            }
-        }
-    };
-    #[derive(serde::Serialize, serde::Deserialize)]
-    #[derive(Default, Debug, PartialEq, Eq, Hash)]
-    /**Custom error with signature `UnknownMigrationConsensusTxSender(uint256,address)` and selector `0x8bd03097`.
-```solidity
-error UnknownMigrationConsensusTxSender(uint256 requestId, address txSender);
-```*/
-    #[allow(non_camel_case_types, non_snake_case, clippy::pub_underscore_fields)]
-    #[derive(Clone)]
-    pub struct UnknownMigrationConsensusTxSender {
-        #[allow(missing_docs)]
-        pub requestId: alloy::sol_types::private::primitives::aliases::U256,
-        #[allow(missing_docs)]
-        pub txSender: alloy::sol_types::private::Address,
-    }
-    #[allow(
-        non_camel_case_types,
-        non_snake_case,
-        clippy::pub_underscore_fields,
-        clippy::style
-    )]
-    const _: () = {
-        use alloy::sol_types as alloy_sol_types;
-        #[doc(hidden)]
-        type UnderlyingSolTuple<'a> = (
-            alloy::sol_types::sol_data::Uint<256>,
-            alloy::sol_types::sol_data::Address,
-        );
-        #[doc(hidden)]
-        type UnderlyingRustTuple<'a> = (
-            alloy::sol_types::private::primitives::aliases::U256,
-            alloy::sol_types::private::Address,
-        );
-        #[cfg(test)]
-        #[allow(dead_code, unreachable_patterns)]
-        fn _type_assertion(
-            _t: alloy_sol_types::private::AssertTypeEq<UnderlyingRustTuple>,
-        ) {
-            match _t {
-                alloy_sol_types::private::AssertTypeEq::<
-                    <UnderlyingSolTuple as alloy_sol_types::SolType>::RustType,
-                >(_) => {}
-            }
-        }
-        #[automatically_derived]
-        #[doc(hidden)]
-        impl ::core::convert::From<UnknownMigrationConsensusTxSender>
-        for UnderlyingRustTuple<'_> {
-            fn from(value: UnknownMigrationConsensusTxSender) -> Self {
-                (value.requestId, value.txSender)
-            }
-        }
-        #[automatically_derived]
-        #[doc(hidden)]
-        impl ::core::convert::From<UnderlyingRustTuple<'_>>
-        for UnknownMigrationConsensusTxSender {
-            fn from(tuple: UnderlyingRustTuple<'_>) -> Self {
-                Self {
-                    requestId: tuple.0,
-                    txSender: tuple.1,
-                }
-            }
-        }
-        #[automatically_derived]
-        impl alloy_sol_types::SolError for UnknownMigrationConsensusTxSender {
-            type Parameters<'a> = UnderlyingSolTuple<'a>;
-            type Token<'a> = <Self::Parameters<
-                'a,
-            > as alloy_sol_types::SolType>::Token<'a>;
-            const SIGNATURE: &'static str = "UnknownMigrationConsensusTxSender(uint256,address)";
-            const SELECTOR: [u8; 4] = [139u8, 208u8, 48u8, 151u8];
-            #[inline]
-            fn new<'a>(
-                tuple: <Self::Parameters<'a> as alloy_sol_types::SolType>::RustType,
-            ) -> Self {
-                tuple.into()
-            }
-            #[inline]
-            fn tokenize(&self) -> Self::Token<'_> {
-                (
-                    <alloy::sol_types::sol_data::Uint<
-                        256,
-                    > as alloy_sol_types::SolType>::tokenize(&self.requestId),
-                    <alloy::sol_types::sol_data::Address as alloy_sol_types::SolType>::tokenize(
-                        &self.txSender,
-                    ),
-                )
-            }
-            #[inline]
-            fn abi_decode_raw_validate(data: &[u8]) -> alloy_sol_types::Result<Self> {
-                <Self::Parameters<
-                    '_,
-                > as alloy_sol_types::SolType>::abi_decode_sequence_validate(data)
-                    .map(Self::new)
-            }
-        }
-    };
-    #[derive(serde::Serialize, serde::Deserialize)]
-    #[derive()]
-    /**Function with signature `initializeFromMigration((uint256,uint256,uint256,uint256,uint256,uint256,(uint8,bytes)[],bytes,address[],bytes32,address[],bytes32,address[],bytes32,uint256,uint8,uint8,uint256))` and selector `0xa0079e0f`.
-```solidity
-function initializeFromMigration(MigrationState memory state) external;
-```*/
-    #[allow(non_camel_case_types, non_snake_case, clippy::pub_underscore_fields)]
-    #[derive(Clone)]
-    pub struct initializeFromMigrationCall {
-        #[allow(missing_docs)]
-        pub state: <MigrationState as alloy::sol_types::SolType>::RustType,
-    }
-    ///Container type for the return parameters of the [`initializeFromMigration((uint256,uint256,uint256,uint256,uint256,uint256,(uint8,bytes)[],bytes,address[],bytes32,address[],bytes32,address[],bytes32,uint256,uint8,uint8,uint256))`](initializeFromMigrationCall) function.
-    #[allow(non_camel_case_types, non_snake_case, clippy::pub_underscore_fields)]
-    #[derive(Clone)]
-    pub struct initializeFromMigrationReturn {}
-    #[allow(
-        non_camel_case_types,
-        non_snake_case,
-        clippy::pub_underscore_fields,
-        clippy::style
-    )]
-    const _: () = {
-        use alloy::sol_types as alloy_sol_types;
-        {
-            #[doc(hidden)]
-            type UnderlyingSolTuple<'a> = (MigrationState,);
-            #[doc(hidden)]
-            type UnderlyingRustTuple<'a> = (
-                <MigrationState as alloy::sol_types::SolType>::RustType,
-            );
-            #[cfg(test)]
-            #[allow(dead_code, unreachable_patterns)]
-            fn _type_assertion(
-                _t: alloy_sol_types::private::AssertTypeEq<UnderlyingRustTuple>,
-            ) {
-                match _t {
-                    alloy_sol_types::private::AssertTypeEq::<
-                        <UnderlyingSolTuple as alloy_sol_types::SolType>::RustType,
-                    >(_) => {}
-                }
-            }
-            #[automatically_derived]
-            #[doc(hidden)]
-            impl ::core::convert::From<initializeFromMigrationCall>
-            for UnderlyingRustTuple<'_> {
-                fn from(value: initializeFromMigrationCall) -> Self {
-                    (value.state,)
-                }
-            }
-            #[automatically_derived]
-            #[doc(hidden)]
-            impl ::core::convert::From<UnderlyingRustTuple<'_>>
-            for initializeFromMigrationCall {
-                fn from(tuple: UnderlyingRustTuple<'_>) -> Self {
-                    Self { state: tuple.0 }
-                }
-            }
-        }
-        {
-            #[doc(hidden)]
-            type UnderlyingSolTuple<'a> = ();
-            #[doc(hidden)]
-            type UnderlyingRustTuple<'a> = ();
-            #[cfg(test)]
-            #[allow(dead_code, unreachable_patterns)]
-            fn _type_assertion(
-                _t: alloy_sol_types::private::AssertTypeEq<UnderlyingRustTuple>,
-            ) {
-                match _t {
-                    alloy_sol_types::private::AssertTypeEq::<
-                        <UnderlyingSolTuple as alloy_sol_types::SolType>::RustType,
-                    >(_) => {}
-                }
-            }
-            #[automatically_derived]
-            #[doc(hidden)]
-            impl ::core::convert::From<initializeFromMigrationReturn>
-            for UnderlyingRustTuple<'_> {
-                fn from(value: initializeFromMigrationReturn) -> Self {
-                    ()
-                }
-            }
-            #[automatically_derived]
-            #[doc(hidden)]
-            impl ::core::convert::From<UnderlyingRustTuple<'_>>
-            for initializeFromMigrationReturn {
-                fn from(tuple: UnderlyingRustTuple<'_>) -> Self {
-                    Self {}
-                }
-            }
-        }
-        impl initializeFromMigrationReturn {
-            fn _tokenize(
-                &self,
-            ) -> <initializeFromMigrationCall as alloy_sol_types::SolCall>::ReturnToken<
-                '_,
-            > {
-                ()
-            }
-        }
-        #[automatically_derived]
-        impl alloy_sol_types::SolCall for initializeFromMigrationCall {
-            type Parameters<'a> = (MigrationState,);
-            type Token<'a> = <Self::Parameters<
-                'a,
-            > as alloy_sol_types::SolType>::Token<'a>;
-            type Return = initializeFromMigrationReturn;
-            type ReturnTuple<'a> = ();
-            type ReturnToken<'a> = <Self::ReturnTuple<
-                'a,
-            > as alloy_sol_types::SolType>::Token<'a>;
-            const SIGNATURE: &'static str = "initializeFromMigration((uint256,uint256,uint256,uint256,uint256,uint256,(uint8,bytes)[],bytes,address[],bytes32,address[],bytes32,address[],bytes32,uint256,uint8,uint8,uint256))";
-            const SELECTOR: [u8; 4] = [160u8, 7u8, 158u8, 15u8];
-            #[inline]
-            fn new<'a>(
-                tuple: <Self::Parameters<'a> as alloy_sol_types::SolType>::RustType,
-            ) -> Self {
-                tuple.into()
-            }
-            #[inline]
-            fn tokenize(&self) -> Self::Token<'_> {
-                (<MigrationState as alloy_sol_types::SolType>::tokenize(&self.state),)
-            }
-            #[inline]
-            fn tokenize_returns(ret: &Self::Return) -> Self::ReturnToken<'_> {
-                initializeFromMigrationReturn::_tokenize(ret)
-            }
-            #[inline]
-            fn abi_decode_returns(data: &[u8]) -> alloy_sol_types::Result<Self::Return> {
-                <Self::ReturnTuple<
-                    '_,
-                > as alloy_sol_types::SolType>::abi_decode_sequence(data)
-                    .map(Into::into)
-            }
-            #[inline]
-            fn abi_decode_returns_validate(
-                data: &[u8],
-            ) -> alloy_sol_types::Result<Self::Return> {
-                <Self::ReturnTuple<
-                    '_,
-                > as alloy_sol_types::SolType>::abi_decode_sequence_validate(data)
-                    .map(Into::into)
-            }
-        }
-    };
-    ///Container for all the [`IKMSGenerationMigration`](self) function calls.
-    #[derive(serde::Serialize, serde::Deserialize)]
-    #[derive()]
-    pub enum IKMSGenerationMigrationCalls {
-        #[allow(missing_docs)]
-        initializeFromMigration(initializeFromMigrationCall),
-    }
-    #[automatically_derived]
-    impl IKMSGenerationMigrationCalls {
-        /// All the selectors of this enum.
-        ///
-        /// Note that the selectors might not be in the same order as the variants.
-        /// No guarantees are made about the order of the selectors.
-        ///
-        /// Prefer using `SolInterface` methods instead.
-        pub const SELECTORS: &'static [[u8; 4usize]] = &[[160u8, 7u8, 158u8, 15u8]];
-    }
-    #[automatically_derived]
-    impl alloy_sol_types::SolInterface for IKMSGenerationMigrationCalls {
-        const NAME: &'static str = "IKMSGenerationMigrationCalls";
-        const MIN_DATA_LENGTH: usize = 736usize;
-        const COUNT: usize = 1usize;
-        #[inline]
-        fn selector(&self) -> [u8; 4] {
-            match self {
-                Self::initializeFromMigration(_) => {
-                    <initializeFromMigrationCall as alloy_sol_types::SolCall>::SELECTOR
-                }
-            }
-        }
-        #[inline]
-        fn selector_at(i: usize) -> ::core::option::Option<[u8; 4]> {
-            Self::SELECTORS.get(i).copied()
-        }
-        #[inline]
-        fn valid_selector(selector: [u8; 4]) -> bool {
-            Self::SELECTORS.binary_search(&selector).is_ok()
-        }
-        #[inline]
-        #[allow(non_snake_case)]
-        fn abi_decode_raw(
-            selector: [u8; 4],
-            data: &[u8],
-        ) -> alloy_sol_types::Result<Self> {
-            static DECODE_SHIMS: &[fn(
-                &[u8],
-            ) -> alloy_sol_types::Result<IKMSGenerationMigrationCalls>] = &[
-                {
-                    fn initializeFromMigration(
-                        data: &[u8],
-                    ) -> alloy_sol_types::Result<IKMSGenerationMigrationCalls> {
-                        <initializeFromMigrationCall as alloy_sol_types::SolCall>::abi_decode_raw(
-                                data,
-                            )
-                            .map(IKMSGenerationMigrationCalls::initializeFromMigration)
-                    }
-                    initializeFromMigration
-                },
-            ];
-            let Ok(idx) = Self::SELECTORS.binary_search(&selector) else {
-                return Err(
-                    alloy_sol_types::Error::unknown_selector(
-                        <Self as alloy_sol_types::SolInterface>::NAME,
-                        selector,
-                    ),
-                );
-            };
-            DECODE_SHIMS[idx](data)
-        }
-        #[inline]
-        #[allow(non_snake_case)]
-        fn abi_decode_raw_validate(
-            selector: [u8; 4],
-            data: &[u8],
-        ) -> alloy_sol_types::Result<Self> {
-            static DECODE_VALIDATE_SHIMS: &[fn(
-                &[u8],
-            ) -> alloy_sol_types::Result<IKMSGenerationMigrationCalls>] = &[
-                {
-                    fn initializeFromMigration(
-                        data: &[u8],
-                    ) -> alloy_sol_types::Result<IKMSGenerationMigrationCalls> {
-                        <initializeFromMigrationCall as alloy_sol_types::SolCall>::abi_decode_raw_validate(
-                                data,
-                            )
-                            .map(IKMSGenerationMigrationCalls::initializeFromMigration)
-                    }
-                    initializeFromMigration
-                },
-            ];
-            let Ok(idx) = Self::SELECTORS.binary_search(&selector) else {
-                return Err(
-                    alloy_sol_types::Error::unknown_selector(
-                        <Self as alloy_sol_types::SolInterface>::NAME,
-                        selector,
-                    ),
-                );
-            };
-            DECODE_VALIDATE_SHIMS[idx](data)
-        }
-        #[inline]
-        fn abi_encoded_size(&self) -> usize {
-            match self {
-                Self::initializeFromMigration(inner) => {
-                    <initializeFromMigrationCall as alloy_sol_types::SolCall>::abi_encoded_size(
-                        inner,
-                    )
-                }
-            }
-        }
-        #[inline]
-        fn abi_encode_raw(&self, out: &mut alloy_sol_types::private::Vec<u8>) {
-            match self {
-                Self::initializeFromMigration(inner) => {
-                    <initializeFromMigrationCall as alloy_sol_types::SolCall>::abi_encode_raw(
-                        inner,
-                        out,
-                    )
-                }
-            }
-        }
-    }
-    ///Container for all the [`IKMSGenerationMigration`](self) custom errors.
-    #[derive(serde::Serialize, serde::Deserialize)]
-    #[derive(Debug, PartialEq, Eq, Hash)]
-    pub enum IKMSGenerationMigrationErrors {
-        #[allow(missing_docs)]
-        InvalidMigrationConsensusState(InvalidMigrationConsensusState),
-        #[allow(missing_docs)]
-        InvalidMigrationCounterState(InvalidMigrationCounterState),
-        #[allow(missing_docs)]
-        InvalidMigrationMaterial(InvalidMigrationMaterial),
-        #[allow(missing_docs)]
-        UnknownMigrationConsensusTxSender(UnknownMigrationConsensusTxSender),
-    }
-    #[automatically_derived]
-    impl IKMSGenerationMigrationErrors {
-        /// All the selectors of this enum.
-        ///
-        /// Note that the selectors might not be in the same order as the variants.
-        /// No guarantees are made about the order of the selectors.
-        ///
-        /// Prefer using `SolInterface` methods instead.
-        pub const SELECTORS: &'static [[u8; 4usize]] = &[
-            [22u8, 187u8, 175u8, 141u8],
-            [69u8, 2u8, 203u8, 241u8],
-            [139u8, 208u8, 48u8, 151u8],
-            [164u8, 211u8, 212u8, 242u8],
-        ];
-    }
-    #[automatically_derived]
-    impl alloy_sol_types::SolInterface for IKMSGenerationMigrationErrors {
-        const NAME: &'static str = "IKMSGenerationMigrationErrors";
-        const MIN_DATA_LENGTH: usize = 0usize;
-        const COUNT: usize = 4usize;
-        #[inline]
-        fn selector(&self) -> [u8; 4] {
-            match self {
-                Self::InvalidMigrationConsensusState(_) => {
-                    <InvalidMigrationConsensusState as alloy_sol_types::SolError>::SELECTOR
-                }
-                Self::InvalidMigrationCounterState(_) => {
-                    <InvalidMigrationCounterState as alloy_sol_types::SolError>::SELECTOR
-                }
-                Self::InvalidMigrationMaterial(_) => {
-                    <InvalidMigrationMaterial as alloy_sol_types::SolError>::SELECTOR
-                }
-                Self::UnknownMigrationConsensusTxSender(_) => {
-                    <UnknownMigrationConsensusTxSender as alloy_sol_types::SolError>::SELECTOR
-                }
-            }
-        }
-        #[inline]
-        fn selector_at(i: usize) -> ::core::option::Option<[u8; 4]> {
-            Self::SELECTORS.get(i).copied()
-        }
-        #[inline]
-        fn valid_selector(selector: [u8; 4]) -> bool {
-            Self::SELECTORS.binary_search(&selector).is_ok()
-        }
-        #[inline]
-        #[allow(non_snake_case)]
-        fn abi_decode_raw(
-            selector: [u8; 4],
-            data: &[u8],
-        ) -> alloy_sol_types::Result<Self> {
-            static DECODE_SHIMS: &[fn(
-                &[u8],
-            ) -> alloy_sol_types::Result<IKMSGenerationMigrationErrors>] = &[
-                {
-                    fn InvalidMigrationMaterial(
-                        data: &[u8],
-                    ) -> alloy_sol_types::Result<IKMSGenerationMigrationErrors> {
-                        <InvalidMigrationMaterial as alloy_sol_types::SolError>::abi_decode_raw(
-                                data,
-                            )
-                            .map(IKMSGenerationMigrationErrors::InvalidMigrationMaterial)
-                    }
-                    InvalidMigrationMaterial
-                },
-                {
-                    fn InvalidMigrationConsensusState(
-                        data: &[u8],
-                    ) -> alloy_sol_types::Result<IKMSGenerationMigrationErrors> {
-                        <InvalidMigrationConsensusState as alloy_sol_types::SolError>::abi_decode_raw(
-                                data,
-                            )
-                            .map(
-                                IKMSGenerationMigrationErrors::InvalidMigrationConsensusState,
-                            )
-                    }
-                    InvalidMigrationConsensusState
-                },
-                {
-                    fn UnknownMigrationConsensusTxSender(
-                        data: &[u8],
-                    ) -> alloy_sol_types::Result<IKMSGenerationMigrationErrors> {
-                        <UnknownMigrationConsensusTxSender as alloy_sol_types::SolError>::abi_decode_raw(
-                                data,
-                            )
-                            .map(
-                                IKMSGenerationMigrationErrors::UnknownMigrationConsensusTxSender,
-                            )
-                    }
-                    UnknownMigrationConsensusTxSender
-                },
-                {
-                    fn InvalidMigrationCounterState(
-                        data: &[u8],
-                    ) -> alloy_sol_types::Result<IKMSGenerationMigrationErrors> {
-                        <InvalidMigrationCounterState as alloy_sol_types::SolError>::abi_decode_raw(
-                                data,
-                            )
-                            .map(
-                                IKMSGenerationMigrationErrors::InvalidMigrationCounterState,
-                            )
-                    }
-                    InvalidMigrationCounterState
-                },
-            ];
-            let Ok(idx) = Self::SELECTORS.binary_search(&selector) else {
-                return Err(
-                    alloy_sol_types::Error::unknown_selector(
-                        <Self as alloy_sol_types::SolInterface>::NAME,
-                        selector,
-                    ),
-                );
-            };
-            DECODE_SHIMS[idx](data)
-        }
-        #[inline]
-        #[allow(non_snake_case)]
-        fn abi_decode_raw_validate(
-            selector: [u8; 4],
-            data: &[u8],
-        ) -> alloy_sol_types::Result<Self> {
-            static DECODE_VALIDATE_SHIMS: &[fn(
-                &[u8],
-            ) -> alloy_sol_types::Result<IKMSGenerationMigrationErrors>] = &[
-                {
-                    fn InvalidMigrationMaterial(
-                        data: &[u8],
-                    ) -> alloy_sol_types::Result<IKMSGenerationMigrationErrors> {
-                        <InvalidMigrationMaterial as alloy_sol_types::SolError>::abi_decode_raw_validate(
-                                data,
-                            )
-                            .map(IKMSGenerationMigrationErrors::InvalidMigrationMaterial)
-                    }
-                    InvalidMigrationMaterial
-                },
-                {
-                    fn InvalidMigrationConsensusState(
-                        data: &[u8],
-                    ) -> alloy_sol_types::Result<IKMSGenerationMigrationErrors> {
-                        <InvalidMigrationConsensusState as alloy_sol_types::SolError>::abi_decode_raw_validate(
-                                data,
-                            )
-                            .map(
-                                IKMSGenerationMigrationErrors::InvalidMigrationConsensusState,
-                            )
-                    }
-                    InvalidMigrationConsensusState
-                },
-                {
-                    fn UnknownMigrationConsensusTxSender(
-                        data: &[u8],
-                    ) -> alloy_sol_types::Result<IKMSGenerationMigrationErrors> {
-                        <UnknownMigrationConsensusTxSender as alloy_sol_types::SolError>::abi_decode_raw_validate(
-                                data,
-                            )
-                            .map(
-                                IKMSGenerationMigrationErrors::UnknownMigrationConsensusTxSender,
-                            )
-                    }
-                    UnknownMigrationConsensusTxSender
-                },
-                {
-                    fn InvalidMigrationCounterState(
-                        data: &[u8],
-                    ) -> alloy_sol_types::Result<IKMSGenerationMigrationErrors> {
-                        <InvalidMigrationCounterState as alloy_sol_types::SolError>::abi_decode_raw_validate(
-                                data,
-                            )
-                            .map(
-                                IKMSGenerationMigrationErrors::InvalidMigrationCounterState,
-                            )
-                    }
-                    InvalidMigrationCounterState
-                },
-            ];
-            let Ok(idx) = Self::SELECTORS.binary_search(&selector) else {
-                return Err(
-                    alloy_sol_types::Error::unknown_selector(
-                        <Self as alloy_sol_types::SolInterface>::NAME,
-                        selector,
-                    ),
-                );
-            };
-            DECODE_VALIDATE_SHIMS[idx](data)
-        }
-        #[inline]
-        fn abi_encoded_size(&self) -> usize {
-            match self {
-                Self::InvalidMigrationConsensusState(inner) => {
-                    <InvalidMigrationConsensusState as alloy_sol_types::SolError>::abi_encoded_size(
-                        inner,
-                    )
-                }
-                Self::InvalidMigrationCounterState(inner) => {
-                    <InvalidMigrationCounterState as alloy_sol_types::SolError>::abi_encoded_size(
-                        inner,
-                    )
-                }
-                Self::InvalidMigrationMaterial(inner) => {
-                    <InvalidMigrationMaterial as alloy_sol_types::SolError>::abi_encoded_size(
-                        inner,
-                    )
-                }
-                Self::UnknownMigrationConsensusTxSender(inner) => {
-                    <UnknownMigrationConsensusTxSender as alloy_sol_types::SolError>::abi_encoded_size(
-                        inner,
-                    )
-                }
-            }
-        }
-        #[inline]
-        fn abi_encode_raw(&self, out: &mut alloy_sol_types::private::Vec<u8>) {
-            match self {
-                Self::InvalidMigrationConsensusState(inner) => {
-                    <InvalidMigrationConsensusState as alloy_sol_types::SolError>::abi_encode_raw(
-                        inner,
-                        out,
-                    )
-                }
-                Self::InvalidMigrationCounterState(inner) => {
-                    <InvalidMigrationCounterState as alloy_sol_types::SolError>::abi_encode_raw(
-                        inner,
-                        out,
-                    )
-                }
-                Self::InvalidMigrationMaterial(inner) => {
-                    <InvalidMigrationMaterial as alloy_sol_types::SolError>::abi_encode_raw(
-                        inner,
-                        out,
-                    )
-                }
-                Self::UnknownMigrationConsensusTxSender(inner) => {
-                    <UnknownMigrationConsensusTxSender as alloy_sol_types::SolError>::abi_encode_raw(
-                        inner,
-                        out,
-                    )
-                }
-            }
-        }
-    }
-    use alloy::contract as alloy_contract;
-    /**Creates a new wrapper around an on-chain [`IKMSGenerationMigration`](self) contract instance.
-
-See the [wrapper's documentation](`IKMSGenerationMigrationInstance`) for more details.*/
-    #[inline]
-    pub const fn new<
-        P: alloy_contract::private::Provider<N>,
-        N: alloy_contract::private::Network,
-    >(
-        address: alloy_sol_types::private::Address,
-        provider: P,
-    ) -> IKMSGenerationMigrationInstance<P, N> {
-        IKMSGenerationMigrationInstance::<P, N>::new(address, provider)
-    }
-    /**Deploys this contract using the given `provider` and constructor arguments, if any.
-
-Returns a new instance of the contract, if the deployment was successful.
-
-For more fine-grained control over the deployment process, use [`deploy_builder`] instead.*/
-    #[inline]
-    pub fn deploy<
-        P: alloy_contract::private::Provider<N>,
-        N: alloy_contract::private::Network,
-    >(
-        provider: P,
-    ) -> impl ::core::future::Future<
-        Output = alloy_contract::Result<IKMSGenerationMigrationInstance<P, N>>,
-    > {
-        IKMSGenerationMigrationInstance::<P, N>::deploy(provider)
-    }
-    /**Creates a `RawCallBuilder` for deploying this contract using the given `provider`
-and constructor arguments, if any.
-
-This is a simple wrapper around creating a `RawCallBuilder` with the data set to
-the bytecode concatenated with the constructor's ABI-encoded arguments.*/
-    #[inline]
-    pub fn deploy_builder<
-        P: alloy_contract::private::Provider<N>,
-        N: alloy_contract::private::Network,
-    >(provider: P) -> alloy_contract::RawCallBuilder<P, N> {
-        IKMSGenerationMigrationInstance::<P, N>::deploy_builder(provider)
-    }
-    /**A [`IKMSGenerationMigration`](self) instance.
-
-Contains type-safe methods for interacting with an on-chain instance of the
-[`IKMSGenerationMigration`](self) contract located at a given `address`, using a given
-provider `P`.
-
-If the contract bytecode is available (see the [`sol!`](alloy_sol_types::sol!)
-documentation on how to provide it), the `deploy` and `deploy_builder` methods can
-be used to deploy a new instance of the contract.
-
-See the [module-level documentation](self) for all the available methods.*/
-    #[derive(Clone)]
-    pub struct IKMSGenerationMigrationInstance<
-        P,
-        N = alloy_contract::private::Ethereum,
-    > {
-        address: alloy_sol_types::private::Address,
-        provider: P,
-        _network: ::core::marker::PhantomData<N>,
-    }
-    #[automatically_derived]
-    impl<P, N> ::core::fmt::Debug for IKMSGenerationMigrationInstance<P, N> {
-        #[inline]
-        fn fmt(&self, f: &mut ::core::fmt::Formatter<'_>) -> ::core::fmt::Result {
-            f.debug_tuple("IKMSGenerationMigrationInstance")
-                .field(&self.address)
-                .finish()
-        }
-    }
-    /// Instantiation and getters/setters.
-    #[automatically_derived]
-    impl<
-        P: alloy_contract::private::Provider<N>,
-        N: alloy_contract::private::Network,
-    > IKMSGenerationMigrationInstance<P, N> {
-        /**Creates a new wrapper around an on-chain [`IKMSGenerationMigration`](self) contract instance.
-
-See the [wrapper's documentation](`IKMSGenerationMigrationInstance`) for more details.*/
-        #[inline]
-        pub const fn new(
-            address: alloy_sol_types::private::Address,
-            provider: P,
-        ) -> Self {
-            Self {
-                address,
-                provider,
-                _network: ::core::marker::PhantomData,
-            }
-        }
-        /**Deploys this contract using the given `provider` and constructor arguments, if any.
-
-Returns a new instance of the contract, if the deployment was successful.
-
-For more fine-grained control over the deployment process, use [`deploy_builder`] instead.*/
-        #[inline]
-        pub async fn deploy(
-            provider: P,
-        ) -> alloy_contract::Result<IKMSGenerationMigrationInstance<P, N>> {
-            let call_builder = Self::deploy_builder(provider);
-            let contract_address = call_builder.deploy().await?;
-            Ok(Self::new(contract_address, call_builder.provider))
-        }
-        /**Creates a `RawCallBuilder` for deploying this contract using the given `provider`
-and constructor arguments, if any.
-
-This is a simple wrapper around creating a `RawCallBuilder` with the data set to
-the bytecode concatenated with the constructor's ABI-encoded arguments.*/
-        #[inline]
-        pub fn deploy_builder(provider: P) -> alloy_contract::RawCallBuilder<P, N> {
-            alloy_contract::RawCallBuilder::new_raw_deploy(
-                provider,
-                ::core::clone::Clone::clone(&BYTECODE),
-            )
-        }
-        /// Returns a reference to the address.
-        #[inline]
-        pub const fn address(&self) -> &alloy_sol_types::private::Address {
-            &self.address
-        }
-        /// Sets the address.
-        #[inline]
-        pub fn set_address(&mut self, address: alloy_sol_types::private::Address) {
-            self.address = address;
-        }
-        /// Sets the address and returns `self`.
-        pub fn at(mut self, address: alloy_sol_types::private::Address) -> Self {
-            self.set_address(address);
-            self
-        }
-        /// Returns a reference to the provider.
-        #[inline]
-        pub const fn provider(&self) -> &P {
-            &self.provider
-        }
-    }
-    impl<P: ::core::clone::Clone, N> IKMSGenerationMigrationInstance<&P, N> {
-        /// Clones the provider and returns a new instance with the cloned provider.
-        #[inline]
-        pub fn with_cloned_provider(self) -> IKMSGenerationMigrationInstance<P, N> {
-            IKMSGenerationMigrationInstance {
-                address: self.address,
-                provider: ::core::clone::Clone::clone(&self.provider),
-                _network: ::core::marker::PhantomData,
-            }
-        }
-    }
-    /// Function calls.
-    #[automatically_derived]
-    impl<
-        P: alloy_contract::private::Provider<N>,
-        N: alloy_contract::private::Network,
-    > IKMSGenerationMigrationInstance<P, N> {
-        /// Creates a new call builder using this contract instance's provider and address.
-        ///
-        /// Note that the call can be any function call, not just those defined in this
-        /// contract. Prefer using the other methods for building type-safe contract calls.
-        pub fn call_builder<C: alloy_sol_types::SolCall>(
-            &self,
-            call: &C,
-        ) -> alloy_contract::SolCallBuilder<&P, C, N> {
-            alloy_contract::SolCallBuilder::new_sol(&self.provider, &self.address, call)
-        }
-        ///Creates a new call builder for the [`initializeFromMigration`] function.
-        pub fn initializeFromMigration(
-            &self,
-            state: <MigrationState as alloy::sol_types::SolType>::RustType,
-        ) -> alloy_contract::SolCallBuilder<&P, initializeFromMigrationCall, N> {
-            self.call_builder(
-                &initializeFromMigrationCall {
-                    state,
-                },
-            )
-        }
-    }
-    /// Event filters.
-    #[automatically_derived]
-    impl<
-        P: alloy_contract::private::Provider<N>,
-        N: alloy_contract::private::Network,
-    > IKMSGenerationMigrationInstance<P, N> {
-        /// Creates a new event filter using this contract instance's provider and address.
-        ///
-        /// Note that the type can be any event, not just those defined in this contract.
-        /// Prefer using the other methods for building type-safe event filters.
-        pub fn event_filter<E: alloy_sol_types::SolEvent>(
-            &self,
-        ) -> alloy_contract::Event<&P, E, N> {
-            alloy_contract::Event::new_sol(&self.provider, &self.address)
-        }
-    }
-}
